requirements.txt
requests
hikari
hikari-lightbulb